WebApp/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.4
django-cors-headers==3.13.0
django-environ==0.9.0
django-smtp-ssl==1.0
djangorestframework==3.14.0
mysqlclient==2.1.1
numpy==1.23.5
pandas==1.5.2
psycopg2==2.9.5
python-dateutil==2.8.2
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
XlsxWriter==3.0.3
